Add automated test-harness for examples and fix DAG tests
