+++ src/hera/workflows/v5/tests/test_workflow.py
@@ -1,58 +1,47 @@
-from hera.workflows.v5.container import Container
-from hera.workflows.v5.workflow import Workflow
+import importlib
+import os
+import pkgutil
+from pathlib import Path
+
+import pytest
+import requests
+import yaml
+
+import hera.workflows.v5.examples as hera_examples
+import hera.workflows.v5.examples.upstream as hera_upstream_examples
+
+ARGO_EXAMPLES_URL = "https://raw.githubusercontent.com/argoproj/argo-workflows/master/examples"
 
 
-def test_workflow_v5():
+@pytest.mark.parametrize(
+    "module_name", [name for _, name, _ in pkgutil.iter_modules(hera_examples.__path__) if name != "upstream"]
+)
+def test_hera_output(module_name):
     # GIVEN
-    with Workflow(
-        generate_name="coinflip-",
-        annotations={
-            "workflows.argoproj.io/description": (
-                "This is an example of coin flip defined as a sequence of conditional steps."
-            ),
-        },
-    ) as w:
-        Container(
-            name="heads",
-            image="alpine:3.6",
-            command=["sh", "-c"],
-            args=["echo 'it was heads'"],
-        )
-        Container(
-            name="tails",
-            image="alpine:3.6",
-            command=["sh", "-c"],
-            args=["echo 'it was tails'"],
-        )
+    workflow = importlib.import_module(f'hera.workflows.v5.examples.{module_name}').w
+    yaml_path = Path(hera_examples.__file__).parent / f"{module_name.replace('_', '-')}.yaml"
     # WHEN
-    output = w.build().to_dict()
+    output = workflow.to_dict()
     # THEN
-    assert output == {
-        "apiVersion": "argoproj.io/v1alpha1",
-        "metadata": {
-            "annotations": {
-                "workflows.argoproj.io/description": "This is an "
-                "example of "
-                "coin flip "
-                "defined as "
-                "a sequence "
-                "of "
-                "conditional "
-                "steps."
-            },
-            "generateName": "coinflip-",
-            "namespace": "default",
-        },
-        "spec": {
-            "templates": [
-                {
-                    "container": {"args": ["echo 'it was heads'"], "command": ["sh", "-c"], "image": "alpine:3.6"},
-                    "name": "heads",
-                },
-                {
-                    "container": {"args": ["echo 'it was tails'"], "command": ["sh", "-c"], "image": "alpine:3.6"},
-                    "name": "tails",
-                },
-            ]
-        },
-    }
+    if not yaml_path.exists() or os.environ.get("HERA_REGENERATE_EXAMPLES"):
+        yaml_path.write_text(yaml.dump(output))
+    assert output == yaml.safe_load(yaml_path.read_text())
+
+
+@pytest.mark.parametrize("module_name", [name for _, name, _ in pkgutil.iter_modules(hera_upstream_examples.__path__)])
+def test_hera_output_upstream(module_name):
+    # GIVEN
+    workflow = importlib.import_module(f'hera.workflows.v5.examples.upstream.{module_name}').w
+    yaml_path = Path(hera_upstream_examples.__file__).parent / f"{module_name.replace('_', '-')}.yaml"
+    upstream_yaml_path = (
+        Path(hera_upstream_examples.__file__).parent / f"upstream.{module_name.replace('_', '-')}.yaml"
+    )
+    # WHEN
+    output = workflow.to_dict()
+    # THEN
+    if not yaml_path.exists() or os.environ.get("HERA_REGENERATE_EXAMPLES"):
+        yaml_path.write_text(yaml.dump(output))
+    if not upstream_yaml_path.exists() or os.environ.get("HERA_REGENERATE_EXAMPLES"):
+        upstream_yaml_path.write_text(requests.get(f"{ARGO_EXAMPLES_URL}/{module_name.replace('_', '-')}.yaml").text)
+    assert output == yaml.safe_load(yaml_path.read_text())
+    assert output == yaml.safe_load(upstream_yaml_path.read_text())
